requirements.txt
aioredis==2.0.1
bcrypt==4.0.1
fastapi==0.100.0
passlib==1.7.4
pydantic==2.0.3
pydantic-settings==2.0.2
python-multipart==0.0.6
pyjwt==2.7.0
starlette==0.27.0
uvicorn==0.23.0
websockets==11.0.3

httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-dotenv== 0.5.2
